Normalize evaluation
With NNUE, our evaluation is no longer related to the classical parameter PawnValueEg (=208).

This leads to the current evaluation changing quite a bit from release to release,
for example, the eval needed to have 50% win rate at fishtest LTC (in cp and internal Value):

June 2020  :   113cp (237)
June 2021  :   115cp (240)
April 2022 :   134cp (279)
July 2022  :   167cp (348)

This inflation can be fixed if one fixes 100cp to mean 50% win chance,
and decouple this conversion from PawnValueEg. This conversion is somewhat arbitrary,
only the relative ranking of positions is important for an engine, which is designed the find the best move.

While there is no simple 1-to-1 relation between the internally used Value, and the win rate,
we can base this on the win_rate_model. The 'a' parameter of this model, gives 50%, and by picking
this value at move 32, this is just the sum of the parameters of the model for a (i.e. the 'as' array).

This patch introduces Internal2Pawn to convert the internal units to cp,
and converts to win_rate_model to internal units.

Generally, it might be better to directly use the wdl values (available with the option UCI_ShowWDL) in analysis,
or focus directly on the bestmove and PV lines provided.

No functional change